AICrawlerLogging/requirements.txt
psycopg2-binary==2.9.10
pandas==2.1.3
python-dateutil==2.8.2
